+++ db/migrate/20210708112144_add_original_report_fk_to_activity.rb
@@ -1,0 +1,7 @@
+class AddOriginalReportFkToActivity < ActiveRecord::Migration[6.1]
+  def change
+    change_table :activities do |t|
+      t.references :originating_report, type: :uuid
+    end
+  end
+end