+++ spec/models/activity_spec.rb
@@ -36,7 +36,8 @@ RSpec.describe Activity, type: :model do
     before do
       allow_any_instance_of(ActivityDefaults).to receive(:call).and_return(
         title: "a returned value",
-        form_state: "form_state"
+        form_state: "form_state",
+        originating_report_id: "fda36c93-794b-4cc4-b15c-2817f00c7c25"
       )
     end
 
@@ -49,6 +50,7 @@ RSpec.describe Activity, type: :model do
       expect(activity).to be_an_instance_of(Activity)
       expect(activity.title).to eq "a returned value"
       expect(activity.form_state).to eq "form_state"
+      expect(activity.originating_report_id).to eq "fda36c93-794b-4cc4-b15c-2817f00c7c25"
     end
 
     it "accepts a block that can override any default values" do
@@ -792,6 +794,7 @@ RSpec.describe Activity, type: :model do
   end
 
   describe "associations" do
+    it { should belong_to(:originating_report).optional }
     it { should belong_to(:organisation) }
     it { should have_many(:child_activities).with_foreign_key("parent_id") }
     it { should belong_to(:extending_organisation).with_foreign_key("extending_organisation_id").optional }